+++ llama-index-integrations/readers/llama-index-readers-web/CHANGELOG.md
@@ -5,4 +5,5 @@
 - Add maintainers and keywords from library.json (llamahub)
 
 ## [0.1.7] - 2024-03-12
-- Add option to WholeSiteReader to pass chrome driver with own options to class __init__
+
+- Add option to WholeSiteReader to pass chrome driver with own options to class **init**